requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.1
filelock==3.14.0
h11==0.14.0
idna==3.7
lxml==5.2.2
ntscraper==0.3.13
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pycparser==2.22
PySocks==1.7.1
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.21.0
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
